api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP (seotracker orchestrator + fallback)
httpx==0.28.1
requests==2.32.3

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Seotracker crawler modules
protego==0.3.1
defusedxml==0.7.1
url-normalize==1.4.3
tldextract==5.1.3
tenacity==9.0.0
openai>=1.30.0
httpx>=0.27.0
playwright>=1.40.0

# Google Search Console + PageSpeed
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
